Modified toApiFormat spec test file to fit standard naming convention. Added tests for code coverage.
