+++ lib/toApiFormat.js
@@ -8,16 +8,17 @@ module.exports = function toApiFormat(source) {
   var excludeList = ['substitution_data', 'tags', 'metadata', 'attributes', 'headers'];
 
   try{
-    // Handle arrays
-    if(Array.isArray(source) && _.isObject(source[0])) {
+    // Handle arrays (Only need to handle arrays of objects for our use case.)
+    if(Array.isArray(source) && _.isPlainObject(source[0])) {
       dest = [];
       for(var i = 0; i < source.length; i++) {
         // Exclude appropriately
-        if( -1 === excludeList.indexOf(source[i])) {
+        /*if( -1 === excludeList.indexOf(source[i])) {
           dest.push(toApiFormat(source[i]));
         } else {
           dest.push(source[i]);
-        }
+        }*/
+        dest.push(toApiFormat(source[i]));
       }
       return dest;
     }
